requirements.txt
fastapi>=0.109.2, <0.110
loguru>=0.7.2, <0.8
rich>=13.7.0, <13.8
uvicorn>=0.27.1, <0.28
pillow>=10.2.0, <10.3
ruff>=0.2.2, <0.3
modal-client>=0.58.83, <0.59
